refactor callstack
