+++ src/variables/description/index.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Jupyter Development Team.
 // Distributed under the terms of the Modified BSD License.
 
-export * from './variableTableDescription';
-export * from './variableDescription';
+export * from './variablesTable';
+export * from './variablesBody';
 export * from './useTbody';
