+++ src/variables/toggle/index.ts
@@ -1,4 +1,4 @@
 // Copyright (c) Jupyter Development Team.
 // Distributed under the terms of the Modified BSD License.
 
-export * from './variableSearch';
+export * from './variablesSearch';
